third_party/WebKit/LayoutTests/css3/calc/catch-divide-by-0-expected.txt

This tests catching of divide by 0 in calc() at parse time

100px / 0 =>
100px / (0) =>
100px / (2 - 2) =>
100px / (2 - (-62 + 64)) =>
100px * (1 / 0) =>
100px * (1 / (0)) =>
100px * (1 / (2 - 2)) =>
100px * (1 / (2 - (-62 + 64))) =>
